cleaned up unused macros and superfluous tests
