+++ packages/weierstrudel/huff_modules/constants.huff
@@ -2,6 +2,8 @@
 #define 2P = takes(0) returns(1) { 0x60C89CE5C263405370A08B6D0302B0BB2F02D522D0E3951A7841182DB0F9FA8E }
 #define 3P = takes(0) returns(1) { 0x912CEB58A394E07D28F0D12384840918C6843FB439555FA7B461A4448976F7D5 }
 #define BETA = takes(0) returns(1) { 0x59e26bcea0d48bacd4f263f1acdb5c4f5763473177fffffe }
-
+#define PRECOMPUTE_ITERATOR = takes(0) returns(0) { 0x60 }
 #define WNAF_MASK = takes(0) returns(1) { 0x1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f1f }
-#define WNAF_OFFSET = takes(0) returns(1) { 0x8b8b8b8b8b8b8b8b8b8b8b8b8b8b8b8b8b8b8b8b8b8b8b8b8b8b8b8b8b8b8b8b }
+#define X2 = takes(0) returns(1) { 0x00 }
+#define Y2 = takes(0) returns(1) { 0x20 }
+#define Z2 = takes(0) returns(1) { 0x40 }
